Divide class NetworkOperations.kt into classes: - RetrofitModule (network package) - NetworkLoad (repository)
