+++ app/src/main/java/ru/aevd/androidacademymovieapp/network/RetrofitModule.kt
@@ -1,0 +1,50 @@
+package ru.aevd.androidacademymovieapp.network
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.Interceptor
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import okhttp3.Response
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import ru.aevd.androidacademymovieapp.BuildConfig
+
+object RetrofitModule {
+    private val json = Json {
+        ignoreUnknownKeys = true
+    }
+
+    private val loggingInterceptor = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BODY
+    }
+
+    //TODO 2: remove logging interceptor in prod
+    private val client = OkHttpClient().newBuilder()
+            .addInterceptor(loggingInterceptor)
+            .addNetworkInterceptor(loggingInterceptor)
+            .addInterceptor(ApiKeyInterceptor())
+            .build()
+
+    @Suppress("EXPERIMENTAL_API_USAGE")
+    private val retrofit: Retrofit = Retrofit.Builder()
+            .client(client)
+            .baseUrl(BuildConfig.BASE_URL)
+            .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+            .build()
+
+    //TODO: you can do it better - watch in the end of the lecture
+    val moviesApi: MoviesApi = retrofit.create(MoviesApi::class.java)
+}
+
+private class ApiKeyInterceptor : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val originalRequest = chain.request()
+        val originalRequestUrl = originalRequest.url
+        val url = originalRequestUrl.newBuilder()
+                .addQueryParameter(BuildConfig.API_KEY_QUERY_PARAM, BuildConfig.API_KEY)
+                .build()
+        val request = originalRequest.newBuilder().url(url).build()
+        return chain.proceed(request = request)
+    }
+}
