+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/fragments/FragmentMoviesList.kt
@@ -12,8 +12,8 @@ import androidx.recyclerview.widget.RecyclerView
 import kotlinx.coroutines.*
 import ru.aevd.androidacademymovieapp.*
 import ru.aevd.androidacademymovieapp.entities.Movie
-import ru.aevd.androidacademymovieapp.network.NetworkOperations
 import ru.aevd.androidacademymovieapp.repository.GetMoviesFromNetwork
+import ru.aevd.androidacademymovieapp.repository.NetworkLoad
 import ru.aevd.androidacademymovieapp.ui.adapters.MoviesAdapter
 import ru.aevd.androidacademymovieapp.ui.adapters.OnMoviesItemClicked
 import ru.aevd.androidacademymovieapp.viewmodels.MoviesListViewModel
@@ -24,7 +24,7 @@ class FragmentMoviesList: Fragment() {
     private val viewModel: MoviesListViewModel by viewModels {
         MoviesListViewModelFactory(
             GetMoviesFromNetwork(
-                NetworkOperations()
+                NetworkLoad()
 //                TODO 4: remove commented old code
 //                GetMoviesFromAssets(
 //                        MoviePersistent(requireContext().applicationContext)
